requirements.txt
beautifulsoup4==4.10.0
Django==3.2.9
line_bot_sdk==2.0.1
opencc==1.1.3
opencc_python_reimplemented==0.1.6
requests==2.24.0
